+++ OAuth/ResourceOwner/GoogleResourceOwner.php
@@ -28,6 +28,9 @@ class GoogleResourceOwner extends GenericOAuth2ResourceOwner
         'infos_url'           => 'https://www.googleapis.com/oauth2/v1/userinfo',
 
         'scope'               => 'https://www.googleapis.com/auth/userinfo.profile',
+
+        // @link https://developers.google.com/accounts/docs/OAuth2WebServer#offline
+        'access_type'         => null,
     );
 
     /**
@@ -40,4 +43,12 @@ class GoogleResourceOwner extends GenericOAuth2ResourceOwner
         'email'          => 'email',
         'profilepicture' => 'picture',
     );
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getAuthorizationUrl($redirectUri, array $extraParameters = array())
+    {
+        return parent::getAuthorizationUrl($redirectUri, array_merge(array('access_type' => $this->getOption('access_type')), $extraParameters));
+    }
 }